refactor: difficulty 관련 클래스를 difficulty 하위 패키지로 이동
- DifficultyBatchService, DifficultySeedCalculator, DynamicDifficultyCalculator, QuoteProfileCalculator를 quote/service/difficulty/ 패키지로 이동
